requirements.txt
argcomplete==2.0.0
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
commitizen==2.36.0
cryptography==38.0.1
decli==0.5.2
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==4.1.2
django-allauth==0.51.0
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.2
packaging==21.3
prompt-toolkit==3.0.31
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python3-openid==3.2.0
pytz==2022.5
PyYAML==6.0
questionary==1.10.0
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.3
termcolor==2.0.1
tomlkit==0.11.6
typing-extensions==4.4.0
tzdata==2022.5
urllib3==1.26.12
wcwidth==0.2.5
